# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Klaw - OpenAPI",
    "description" : "This specification is still a work in progress and is not yet implemented in any API. The purpose of this specification is to facilitate developers discussions.",
    "contact" : {
      "email" : "info@klaw-project.io"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "2.10.0"
  },
...
  "servers" : [ {
    "url" : "http://localhost:9097",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/opReq/reqId/{reqId}/delete" : {
      "post" : {
        "tags" : [ "operational-requests-controller" ],
        "operationId" : "deleteOperationalRequest",
        "parameters" : [ {
          "name" : "reqId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/opReq/reqId/{reqId}/decline" : {
      "post" : {
        "tags" : [ "operational-requests-controller" ],
        "operationId" : "declineOperationalRequest",
        "parameters" : [ {
          "name" : "reqId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "reasonForDecline",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/opReq/reqId/{reqId}/approve" : {
      "post" : {
        "tags" : [ "operational-requests-controller" ],
        "operationId" : "approveOperationalRequest",
        "parameters" : [ {
          "name" : "reqId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/opReq/consumerOffsetsReset/create" : {
      "post" : {
        "tags" : [ "operational-requests-controller" ],
        "operationId" : "createConsumerOffsetsResetRequest",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConsumerOffsetResetRequestModel"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/topic/request/{topicReqId}" : {
      "get" : {
        "tags" : [ "topic-controller" ],
        "operationId" : "getTopicRequest",
        "parameters" : [ {
          "name" : "topicReqId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopicRequestsResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/schema/request/{schemaReqId}" : {
      "get" : {
        "tags" : [ "schema-registry-controller" ],
        "operationId" : "getSchemaRequest",
        "parameters" : [ {
          "name" : "schemaReqId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchemaRequestsResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/operationalRequests/requestsFor/{requestsFor}" : {
      "get" : {
        "tags" : [ "operational-requests-controller" ],
        "operationId" : "getOperationalRequests",
        "parameters" : [ {
          "name" : "requestsFor",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "pageNo",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "currentPage",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : ""
          }
        }, {
          "name" : "requestStatus",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "CREATED", "DELETED", "DECLINED", "APPROVED", "ALL" ]
          }
        }, {
          "name" : "env",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "topicName",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "consumerGroup",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "operationType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "RESET_CONSUMER_OFFSETS" ]
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "order",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "DESC_REQUESTED_TIME",
            "enum" : [ "ASC_REQUESTED_TIME", "DESC_REQUESTED_TIME" ]
          }
        }, {
          "name" : "isMyRequest",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/OperationalRequestsResponseModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/opReq/consumerOffsetsReset/validate" : {
      "get" : {
        "tags" : [ "operational-requests-controller" ],
        "operationId" : "validateOffsetRequestDetails",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "topicName",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "consumerGroup",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvIdInfo"
                }
              }
            }
          }
        }
      }
    },
...
    "/acl/request/{aclRequestId}" : {
      "get" : {
        "tags" : [ "acl-controller" ],
        "operationId" : "getAclRequest",
        "parameters" : [ {
          "name" : "aclRequestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AclRequestsResponseModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponse" : {
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "errCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "debugMessage" : {
            "type" : "string"
          },
          "data" : { },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "message", "success" ]
      },
...
      "TopicConfigEntry" : {
        "properties" : {
          "configKey" : {
            "type" : "string"
          },
          "configValue" : {
            "type" : "string"
          }
        }
      },
...
      "EnvIdInfo" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
...
      "ConsumerOffsetResetRequestModel" : {
        "properties" : {
          "operationalRequestType" : {
            "type" : "string",
            "enum" : [ "RESET_CONSUMER_OFFSETS" ]
          },
          "environment" : {
            "type" : "string"
          },
          "requestingTeamId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "approvingTeamId" : {
            "type" : "string"
          },
          "otherParams" : {
            "type" : "string"
          },
          "appname" : {
            "type" : "string"
          },
          "remarks" : {
            "type" : "string"
          },
          "requestor" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "topicname" : {
            "type" : "string",
            "pattern" : "^[a-zA-Z0-9._-]{3,}$"
          },
          "consumerGroup" : {
            "type" : "string",
            "pattern" : "^[a-zA-Z0-9._-]{3,}$"
          },
          "offsetResetType" : {
            "type" : "string",
            "enum" : [ "LATEST", "EARLIEST", "TO_DATE_TIME" ]
          },
          "resetTimeStampStr" : {
            "type" : "string"
          }
        },
        "required" : [ "consumerGroup", "environment", "offsetResetType", "operationalRequestType", "topicname" ]
      },
...
      "AclApproval" : {
        "properties" : {
          "approvalId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "approvalType" : {
            "type" : "string",
            "enum" : [ "TOPIC_TEAM_OWNER", "CONNECTOR_TEAM_OWNER", "ACL_TEAM_OWNER", "TEAM" ]
          },
          "requiredApprover" : {
            "type" : "string"
          },
          "approverName" : {
            "type" : "string"
          },
          "approverTeamId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "approverTeamName" : {
            "type" : "string"
          },
          "parent" : {
            "$ref" : "#/components/schemas/AclRequests"
          }
        }
      },
      "AclRequests" : {
        "properties" : {
          "req_no" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tenantId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "topicname" : {
            "type" : "string"
          },
          "environment" : {
            "type" : "string"
          },
          "environmentName" : {
            "type" : "string"
          },
          "teamId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "requestingteam" : {
            "type" : "integer",
            "format" : "int32"
          },
          "appname" : {
            "type" : "string"
          },
          "aclType" : {
            "type" : "string"
          },
          "consumergroup" : {
            "type" : "string"
          },
          "requestor" : {
            "type" : "string"
          },
          "requesttime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requesttimestring" : {
            "type" : "string"
          },
          "requestStatus" : {
            "type" : "string"
          },
          "remarks" : {
            "type" : "string"
          },
          "acl_ip" : {
            "type" : "string"
          },
          "acl_ssl" : {
            "type" : "string"
          },
          "approver" : {
            "type" : "string"
          },
          "approvingtime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requestOperationType" : {
            "type" : "string"
          },
          "aclPatternType" : {
            "type" : "string"
          },
          "aclResourceType" : {
            "type" : "string"
          },
          "transactionalId" : {
            "type" : "string"
          },
          "otherParams" : {
            "type" : "string"
          },
          "jsonParams" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "aclIpPrincipleType" : {
            "type" : "string",
            "enum" : [ "IP_ADDRESS", "PRINCIPAL", "USERNAME" ]
          },
          "associatedAclId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalNoPages" : {
            "type" : "string"
          },
          "currentPage" : {
            "type" : "string"
          },
          "allPageNos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "TopicRequestsResponseModel" : {
        "properties" : {
          "environment" : {
            "type" : "string"
          },
          "environmentName" : {
            "type" : "string"
          },
          "requestor" : {
            "type" : "string"
          },
          "teamId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teamname" : {
            "type" : "string"
          },
          "requestOperationType" : {
            "type" : "string",
            "enum" : [ "CREATE", "UPDATE", "PROMOTE", "CLAIM", "DELETE", "ALL" ]
          },
          "requestStatus" : {
            "type" : "string",
            "enum" : [ "CREATED", "DELETED", "DECLINED", "APPROVED", "ALL" ]
          },
          "requesttime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requesttimestring" : {
            "type" : "string"
          },
          "currentPage" : {
            "type" : "string"
          },
          "totalNoPages" : {
            "type" : "string"
          },
          "allPageNos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "approvingTeamDetails" : {
            "type" : "string"
          },
          "approver" : {
            "type" : "string"
          },
          "approvingtime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "remarks" : {
            "type" : "string"
          },
          "appname" : {
            "type" : "string"
          },
          "otherParams" : {
            "type" : "string"
          },
          "topicname" : {
            "type" : "string"
          },
          "topicpartitions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "replicationfactor" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "topicid" : {
            "type" : "integer",
            "format" : "int32"
          },
          "deleteAssociatedSchema" : {
            "type" : "boolean"
          },
          "advancedTopicConfigEntries" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TopicConfigEntry"
            }
          },
          "approvingTeamId" : {
            "type" : "string"
          },
          "sequence" : {
            "type" : "string"
          },
          "possibleTeams" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "deletable" : {
            "type" : "boolean"
          },
          "editable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "advancedTopicConfigEntries", "allPageNos", "approvingTeamDetails", "currentPage", "deleteAssociatedSchema", "description", "environment", "environmentName", "replicationfactor", "requestOperationType", "requestStatus", "requestor", "requesttime", "requesttimestring", "teamId", "teamname", "topicid", "topicname", "topicpartitions", "totalNoPages" ]
      },
...
      "SchemaRequestsResponseModel" : {
        "properties" : {
          "environment" : {
            "type" : "string"
          },
          "environmentName" : {
            "type" : "string"
          },
          "requestor" : {
            "type" : "string"
          },
          "teamId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teamname" : {
            "type" : "string"
          },
          "requestOperationType" : {
            "type" : "string",
            "enum" : [ "CREATE", "UPDATE", "PROMOTE", "CLAIM", "DELETE", "ALL" ]
          },
          "requestStatus" : {
            "type" : "string",
            "enum" : [ "CREATED", "DELETED", "DECLINED", "APPROVED", "ALL" ]
          },
          "requesttime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requesttimestring" : {
            "type" : "string"
          },
          "currentPage" : {
            "type" : "string"
          },
          "totalNoPages" : {
            "type" : "string"
          },
          "allPageNos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "approvingTeamDetails" : {
            "type" : "string"
          },
          "approver" : {
            "type" : "string"
          },
          "approvingtime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "remarks" : {
            "type" : "string"
          },
          "appname" : {
            "type" : "string"
          },
          "otherParams" : {
            "type" : "string"
          },
          "topicname" : {
            "type" : "string"
          },
          "schemafull" : {
            "type" : "string"
          },
          "req_no" : {
            "type" : "integer",
            "format" : "int32"
          },
          "forceRegister" : {
            "type" : "boolean"
          },
          "schemaType" : {
            "type" : "string",
            "enum" : [ "AVRO", "JSON" ]
          },
          "schemaversion" : {
            "type" : "string"
          },
          "deletable" : {
            "type" : "boolean"
          },
          "editable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "allPageNos", "approvingTeamDetails", "currentPage", "environment", "environmentName", "forceRegister", "req_no", "requestOperationType", "requestStatus", "requestor", "requesttime", "requesttimestring", "schemaType", "schemafull", "teamId", "teamname", "topicname", "totalNoPages" ]
      },
...
      "OperationalRequestsResponseModel" : {
        "properties" : {
          "topicname" : {
            "type" : "string"
          },
          "consumerGroup" : {
            "type" : "string"
          },
          "offsetResetType" : {
            "type" : "string",
            "enum" : [ "LATEST", "EARLIEST", "TO_DATE_TIME" ]
          },
          "resetTimeStampStr" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "reqId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "environment" : {
            "type" : "string"
          },
          "environmentName" : {
            "type" : "string"
          },
          "requestor" : {
            "type" : "string"
          },
          "teamId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teamname" : {
            "type" : "string"
          },
          "operationalRequestType" : {
            "type" : "string",
            "enum" : [ "RESET_CONSUMER_OFFSETS" ]
          },
          "requestStatus" : {
            "type" : "string",
            "enum" : [ "CREATED", "DELETED", "DECLINED", "APPROVED", "ALL" ]
          },
          "requesttime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requesttimestring" : {
            "type" : "string"
          },
          "currentPage" : {
            "type" : "string"
          },
          "totalNoPages" : {
            "type" : "string"
          },
          "allPageNos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "approvingTeamDetails" : {
            "type" : "string"
          },
          "approver" : {
            "type" : "string"
          },
          "approvingtime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "remarks" : {
            "type" : "string"
          },
          "appname" : {
            "type" : "string"
          },
          "otherParams" : {
            "type" : "string"
          },
          "approvingTeamId" : {
            "type" : "string"
          },
          "sequence" : {
            "type" : "string"
          },
          "possibleTeams" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "deletable" : {
            "type" : "boolean"
          },
          "editable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "allPageNos", "approvingTeamDetails", "consumerGroup", "currentPage", "description", "environment", "environmentName", "offsetResetType", "operationalRequestType", "reqId", "requestStatus", "requestor", "requesttime", "requesttimestring", "teamId", "teamname", "topicname", "totalNoPages" ]
      },
...
      "AclRequestsResponseModel" : {
        "properties" : {
          "environment" : {
            "type" : "string"
          },
          "environmentName" : {
            "type" : "string"
          },
          "requestor" : {
            "type" : "string"
          },
          "teamId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "teamname" : {
            "type" : "string"
          },
          "requestOperationType" : {
            "type" : "string",
            "enum" : [ "CREATE", "UPDATE", "PROMOTE", "CLAIM", "DELETE", "ALL" ]
          },
          "requestStatus" : {
            "type" : "string",
            "enum" : [ "CREATED", "DELETED", "DECLINED", "APPROVED", "ALL" ]
          },
          "requesttime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requesttimestring" : {
            "type" : "string"
          },
          "currentPage" : {
            "type" : "string"
          },
          "totalNoPages" : {
            "type" : "string"
          },
          "allPageNos" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "approvingTeamDetails" : {
            "type" : "string"
          },
          "approver" : {
            "type" : "string"
          },
          "approvingtime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "remarks" : {
            "type" : "string"
          },
          "appname" : {
            "type" : "string"
          },
          "otherParams" : {
            "type" : "string"
          },
          "topicname" : {
            "type" : "string"
          },
          "aclPatternType" : {
            "type" : "string"
          },
          "aclType" : {
            "type" : "string",
            "enum" : [ "PRODUCER", "CONSUMER" ]
          },
          "aclIpPrincipleType" : {
            "type" : "string",
            "enum" : [ "IP_ADDRESS", "PRINCIPAL", "USERNAME" ]
          },
          "req_no" : {
            "type" : "integer",
            "format" : "int32"
          },
          "requestingteam" : {
            "type" : "integer",
            "format" : "int32"
          },
          "requestingTeamName" : {
            "type" : "string"
          },
          "consumergroup" : {
            "type" : "string"
          },
          "acl_ip" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "acl_ssl" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "approvals" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AclApproval"
            }
          },
          "transactionalId" : {
            "type" : "string"
          },
          "aclResourceType" : {
            "type" : "string"
          },
          "deletable" : {
            "type" : "boolean"
          },
          "editable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "aclIpPrincipleType", "aclPatternType", "aclType", "acl_ip", "acl_ssl", "allPageNos", "approvingTeamDetails", "consumergroup", "currentPage", "environment", "environmentName", "req_no", "requestOperationType", "requestStatus", "requestingTeamName", "requestingteam", "requestor", "requesttime", "requesttimestring", "teamId", "teamname", "topicname", "totalNoPages" ]
      }
    }
  }
}